% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getShoppyInvItem.R
\name{getShoppyInvItem}
\alias{getShoppyInvItem}
\title{Shopify Inventory Items Endpoint (Doesn't work)}
\usage{
getShoppyInvItem(shopifyPath, apiKey, apiPassword, verbose = T, page = NULL)
}
\arguments{
\item{shopifyPath}{Something like: mywebsite.com}

\item{apiKey}{Unencoded api key}

\item{apiPassword}{Unencoded apiPassword}

\item{verbose}{Whether it will return the results of the api call. Defaults to T.}

\item{page}{Returns a specific page of results. Defaults to 1.}
}
\value{
Returns a dataframe of inventory items. May contain nested dataframes.
}
\description{
Gets shopify inventory items

Documentation: https://help.shopify.com/api/reference/inventory/inventoryitem
}
