% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getShopifyLocations.R
\name{getShoppyLocations}
\alias{getShoppyLocations}
\title{Shopify Location Endpoint}
\usage{
getShoppyLocations(shopifyPath, apiKey, apiPassword, verbose = T,
  locationId = NULL)
}
\arguments{
\item{shopifyPath}{Something like: mywebsite.com}

\item{apiKey}{Unencoded api key}

\item{apiPassword}{Unencoded apiPassword}

\item{verbose}{Whether it will return the results of the api call. Defaults to T.}

\item{locationID}{Id of a specific location in quotes.}
}
\value{
If locationId is missing, then it returns a dataframe about each location. Otherwise, it returns html about a specific location.
}
\description{
Gets shopify products.

Documentation: https://help.shopify.com/api/reference/location#index
}
