% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getShoppyOrders.R
\name{getShoppyOrders}
\alias{getShoppyOrders}
\title{Shopify Orders API endpoint}
\usage{
getShoppyOrders(shopifyPath, apiKey, apiPassword, verbose = T)
}
\arguments{
\item{shopifyPath}{Something like: mywebsite.com}

\item{apiKey}{Unencoded api key}

\item{apiPassword}{Unencoded apiPassword}

\item{verbose}{Whether it will return the results of the api call. Defaults to T.}
}
\value{
A dataframe of orders. This dataframe contains nested values.
}
\description{
Gets shopify orders by hour.

Documentation: https://help.shopify.com/api/reference/orders/order#index

Requires the data to already be summarized into daily amounts. Make sure that there are no missing periods of coviate data if covariates are being included in the model.
}
