% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getShoppyProducts.R
\name{getShoppyProducts}
\alias{getShoppyProducts}
\title{Shopify Products Endpoint}
\usage{
getShoppyProducts(shopifyPath, apiKey, apiPassword, verbose = T,
  createdMin = NULL, createdMax = NULL, page = NULL)
}
\arguments{
\item{shopifyPath}{Something like: mywebsite.com}

\item{apiKey}{Unencoded api key}

\item{apiPassword}{Unencoded apiPassword}

\item{verbose}{Whether it will return the results of the api call. Defaults to T.}

\item{createdMin}{Show orders created at or after date (format: 2014-04-25T16:15:47-04:00).}

\item{createdMax}{Show orders created at or before date (format: 2014-04-25T16:15:47-04:00).}

\item{page}{Returns a specific page of results. Defaults to 1.}
}
\value{
Returns a dataframe of products. May contain nested dataframes.
}
\description{
Gets shopify products.

Documentation: https://help.shopify.com/api/reference/products#index
}
