% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getShoppyOrders.R
\name{getShoppyOrders}
\alias{getShoppyOrders}
\title{Shopify Orders API endpoint}
\usage{
getShoppyOrders(shopifyPath, apiKey, apiPassword, count = F, verbose = T,
  createdMin = NULL, createdMax = NULL)
}
\arguments{
\item{shopifyPath}{Something like: mywebsite.com}

\item{apiKey}{Unencoded api key}

\item{apiPassword}{Unencoded apiPassword}

\item{count}{Defaults to False. If set to True, it will only return the count of orders.}

\item{verbose}{Whether it will return the results of the api call. Defaults to T.}

\item{createdMin}{Show orders created at or after date (format: 2014-04-25T16:15:47-04:00).}

\item{createdMax}{Show orders created at or before date (format: 2014-04-25T16:15:47-04:00).}
}
\value{
If count == T, then a number. Otherwise, a dataframe of orders. This dataframe may contain nested values.
}
\description{
Gets shopify orders count, all orders, or orders by hour. Returns a max of 250 orders.

Documentation: https://help.shopify.com/api/reference/orders/order#index
}
